* 'main' of https://github.com/ikohong88/Team6_BE:
  - 테스트용 예외처리 추가
  댓글 작성 버그 수정

+++ src/main/java/com/example/troubleshooter/controller/CommentController.java
@@ -19,18 +19,21 @@ public class CommentController {
 
     @PostMapping("/api/posts/{postId}/comments")
     public ResponseEntity postComment(@PathVariable Long postId, @RequestBody @Valid CommentRequestDto commentRequestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) throw new IllegalArgumentException("로그인이 필요합니다.");
         commentService.writeComment(postId, commentRequestDto, userDetails.getUser().getId());
         return new ResponseEntity(HttpStatus.CREATED);
     }
 
     @PutMapping("/api/comments/{commentId}")
     public ResponseEntity putComment(@PathVariable Long commentId, @RequestBody @Valid CommentRequestDto commentRequestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) throw new IllegalArgumentException("로그인이 필요합니다.");
         commentService.editComment(commentId, commentRequestDto, userDetails.getUser().getId());
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @DeleteMapping("/api/comments/{commentId}")
     public ResponseEntity deleteComment(@PathVariable Long commentId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) throw new IllegalArgumentException("로그인이 필요합니다.");
         commentService.deleteComment(commentId, userDetails.getUser().getId());
         return new ResponseEntity(HttpStatus.OK);
     }
